+++ src/components/header/index.tsx
@@ -1,4 +1,4 @@
-import { Contact } from "@components/contact";
+import { NavigationLink } from "@components/menu";
 import Link from "next/link";
 import React from "react";
 
@@ -8,7 +8,7 @@ export const Header: React.FC = () => {
             className="
         w-full fixed flex flex-row justify-between items-center
         h-16 md:h-20 z-50
-        px-8 md:px-16 lg:px-32 xl:px-56 2xl:px-72
+        px-4 md:px-8 lg:px-16 xl:px-28 2xl:px-36
         "
             data-testid="container"
         >
@@ -16,7 +16,12 @@ export const Header: React.FC = () => {
                 <a className="logo">AJ</a>
             </Link>
             <div className="flex flex-row items-center">
-                <Contact />
+                <NavigationLink text="About" href="#about" />
+                <NavigationLink text="Blog" href="/blog" />
+                <NavigationLink
+                    text="Photography"
+                    href="https://alanjohn.myportfolio.com/"
+                />
             </div>
         </div>
     );