+++ src/components/section/index.tsx
@@ -2,12 +2,21 @@ import { ReactNode } from "react";
 
 type Props = {
     children: ReactNode;
+    title?: string;
+    className?: string;
 };
 
-export const Section: React.FC<Props> = ({ children }: Props) => {
+export const Section: React.FC<Props> = ({
+    children,
+    title = "",
+    className = "",
+}: Props) => {
     return (
-        <section className="flex flex-col w-full min-h-screen justify-center">
-            { children }
+        <section
+            id={title}
+            className={`flex flex-col w-full ${className}`}
+        >
+            {children}
         </section>
     );
 };