+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.2.2] - 2022-01-28
+### Added
+- Button components
+- Navigation links
+- About me section with filler text
+- Image component
+- Release branch automation using actions
+### Fixed
+- Section to be customizable at page level
+- Header structure
+
 ### [0.2.1] - 2022-01-28
 ### Removed
 - Animated background due to performance issues